fs: move NoDirectoriesMixin to http filesystem
